Update dllnotfoundexception.md
Spent a day on debugging this problem which in the end turned out to be just some problem with the linking of my native library. Not 100% sure if the description is fully correct as I'm not really up to speed on how native dependencies are resolved. This however is the manner in which I found my problem.