From Mono 4.2 will be supported
There are a bunch of fixes in Mono 4.2 which deals with MVC 5 support. Currently have MVC 5 running on mono with exception to the fixes I'm waiting for. The only thing that was not working was the lowercase URL's and the append tailing slash function was missing. This should now be fixed.