Corrected reference to tlstest (mono-basics.md) because it's not working in master
Since tlstest was removed in master, I replaced the link to tlstest with an older branch.
I have also corrected the command to compile it, because it needs a reference to Mono.Security.dll otherwise we get this error: 
tlstest.cs(23,21): error CS0234: The type or namespace name 'Protocol' does not exist in the namespace 'Mono.Security'. Are you missing an assembly reference?