Clean up x86 page
I've made a few formatting changes, but I've also removed the statement "Instruction selector is able to pick best instructions on x86". I don't think it makes sense as an "optimization" given that it should be a given that the compiler chooses instructions from the correct instruction set. I could have misinterpreted that though, so I'm happy to add it back if need be.