Ensures .gnupg directory exists for root
gnupg is quite finnicky when retrieving keys from a keyserver to a standalone keyring. 

A .gnupg directory *must* exist (preferably with correct permission) for the invoking user to temporarily store a keyring.

Since keys used for apt distribution are recommended to be stored in a root-accessible only directory, gpg is invoked by the root user, and /root/.gnupg must be created (as it typically does not exist). This is entirely harmless but necessary to avoid gpg erroring on inexistant directory.